Moved Joda and Android specifics into separate project
Joda Time is deprecated in favor of Java Time, so there is not much point in keeping it in main library.
Android specifics are not used that often so move them to separate project to get rid of all provided dependencies.
Moved Date/Timestamp tests to Java8/Joda projects

TODO: reimport Timestamp tests for Java8