Preparing v1.7.2
When non empty ctor is used, array index or is not mandatory
Allow duplicate properties as long as annotation is not used
Change Scala API so TypeTag is not mandatory. TypeTag allocates and thus when used in loop library seems slow.
Add enum to gen-code.
Add support for Jsonb (high level only)