Fix broken JsonObject behavior
Ideally converters should support specialized JsonObject, but as an alternative just have wrapper to JsonObject behavior from the standard API.
This introduces some minor overhead, but it removes some overhead on other places.

Java8 processor will create specialized behavior for JsonObject, but if they are used in containers, runtime conversion needs to kick in and handle it.