Improve analysis warnings. Unify bean naming
When annotation is put on non processed field/method, produce a warning with explanation.
Ideally we could produce error, but this would require differentiation with other annotation types or rewriting tools, so for now just produce a warning which should be good enough.

On bean methods, detect a special case where all characters are uppercased. In that case the name should be all uppercased also.
Merge various bean naming implementation into a single place for easier further changes.