Add @Generated annotation to generated code
The @Generated annotation is used by many 3rd party tools to flag code that should be handled differently. In particular, it is a relatively standardised way to signal that code should be excluded from automated analysis.

One example is the NullAway annotation processor. This checks for potential nullability errors, but this is unnecessary in the generated dsl-json code.