Minor bugfixes
Cope with java.sql.Date as it cannot be converted toInstant and instead use special method to handle it.
Fix bug with key parsing as it was handling double quoting incorrectly.
While originally this was fine, when names could only be regular properties, now it's not fine anymore so cope with it.