Support no setter collections
When collection is explicitly marked with an annotation, does not have setter and is non nullable use appropriate pattern to process it.
Assume it can't be null and append elements to it

While it would be better to use collection instance directly instead of allocating new collection and then copying, for now just reuse existing infrastructure.

This resolves the problem of not recognizing such collections even when explicitly marked.
Emit warnings on missing NonNull markers

Add dsl-json NonNull marker as supported one.

Introduce basic infrastructure for reading/writing separation. This kind of works already for some cases... it should work in even more cases in the future.

Mark most public methods on Analysis as private to avoid breaking changes due to minor improvements.