Support for default values
JsonConverter now supports definition of default.
jsonDefault() method or JSON_DEFAULT field can be used.
When defined field will be initialized with such value, so if missing from JSON it will end up as such even if default constructor has different value.

When serializing with omit it will be compared (instance based) for equality.
This means it can be used on primitives or reused instances, but not really on all objects, at least for serialization.